// src/app/Interfaces/partner.interface.ts
export interface Partner{
    id: number;
    code: string;
    active: string;
    fullname: string;
    cpf_cnpj: number;
    age: number;
    type: string;
    email: string;
    phone1: number;
    phone2: number;
    cep: number;
    address: string;
    streetnum: number;
    city: string;
    block: string;
    uf: string;
}